Don't fail on missing container (#80)
This will solve a MenuBar issue https://github.com/vaadin/vaadin-menu-bar-flow/issues/29
When MenuBar sets a root item hidden, the item will have a
_containerNodeId (like it should when it has children), but that
container is not actually found from Flow-managed components in the
client-side. This happens for an unknown reason, which I couldn't solve.
But I've verified that this fix works fine. The item doesn't need to
have children when it is hidden, as they are re-generated again when the
item is set back to visible.

This will be tested in MenuBar: https://github.com/vaadin/vaadin-menu-bar-flow/pull/37