Sync checked-state to client when attached but not yet opened (#83)
This is for https://github.com/vaadin/vaadin-menu-bar-flow/issues/46.
The issue will be closed by another PR which adds the test for MenuBar.

The same problem applied also for a ContextMenu which was manually
attached to DOM, but not yet opened when setting an item checked with
the Java API.

It was caused by false assumption that as soon as the items-array is set
for the menu, each component in there will have _item reference. The
_item reference is however added by the web component when the
item-component is rendered. This allowed the checked state to go out of
sync if the items-array is already generated (on attach, this will also
sync the checked state), but before the item-component is rendered (so
it doesn't have the _item reference yet).

The fix was to add the _item reference already while generating the
items-array for the web component.